Organize tests to mirror package structure
