+++ README.md
@@ -241,6 +241,7 @@ docker run --rm -v C:\path\to\config.yml:/app/config.yml \
   (русская версия доступна в `AGENTS.ru.md`).
 - Убедитесь, что новые скрипты содержат информативные docstring и защищённую точку входа `main`.
 - Перед коммитом выполните `python -m compileall -q .` для проверки синтаксиса.
+- Тесты в каталоге `tests/` повторяют структуру пакета `finmodel` (например, `tests/utils/`).
 
 ## Логи
 Все скрипты пишут логи в файл `log/finmodel.log`. Создайте каталог `log/`, если его
