+++ tests/AGENTS.md
@@ -1,0 +1,4 @@
+# Tests Instructions
+
+- Mirror the `finmodel` package structure within this directory.
+- Place tests for `finmodel.utils` under `tests/utils/`, for `finmodel.scripts` under `tests/scripts/`, and so on.
